Move common classes into their own packages
